+++ src/Configuration.Writable.Core/Options/WritableOptionsImpl.cs
@@ -154,16 +154,6 @@ internal sealed class WritableOptionsImpl<T>(
     IReadOnlyOptions<T> IReadOnlyNamedOptions<T>.GetInstance(string name) =>
         new WritableOptionsWithNameImpl<T>(this, name);
 
-    /// <inheritdoc />
-    [System.Obsolete("Use GetInstance instead.")]
-    IWritableOptions<T> IWritableNamedOptions<T>.GetSpecifiedInstance(string name) =>
-        ((IWritableNamedOptions<T>)this).GetInstance(name);
-
-    /// <inheritdoc />
-    [System.Obsolete("Use GetInstance instead.")]
-    IReadOnlyOptions<T> IReadOnlyNamedOptions<T>.GetSpecifiedInstance(string name) =>
-        ((IReadOnlyNamedOptions<T>)this).GetInstance(name);
-
     /// <summary>
     /// Asynchronously saves the specified configuration.
     /// </summary>