+++ src/Configuration.Writable.Core/Testing/WritableOptionsStub.cs
@@ -181,16 +181,6 @@ public class WritableOptionsStub<T> : IWritableOptionsMonitor<T>
     IReadOnlyOptions<T> IReadOnlyNamedOptions<T>.GetInstance(string name) =>
         new WritableOptionsStubWithName<T>(this, name);
 
-    /// <inheritdoc/>
-    [System.Obsolete("Use GetInstance instead.")]
-    IWritableOptions<T> IWritableNamedOptions<T>.GetSpecifiedInstance(string name) =>
-        ((IWritableNamedOptions<T>)this).GetInstance(name);
-
-    /// <inheritdoc/>
-    [System.Obsolete("Use GetInstance instead.")]
-    IReadOnlyOptions<T> IReadOnlyNamedOptions<T>.GetSpecifiedInstance(string name) =>
-        ((IReadOnlyNamedOptions<T>)this).GetInstance(name);
-
     // A simple disposable action implementation
     private sealed class DisposableAction(Action disposeAction) : IDisposable
     {